requirements.txt
azure-functions
azure-functions-durable
azure-identity
azure-servicebus
pyodbc
sqlalchemy
google-api-python-client
google-auth-oauthlib
google-auth
azure-ai-projects
azure-ai-agents
jinja2
